Refactor DemandController and DemandHandler
