+++ lib/membrane/core/element/action_handler.ex
@@ -315,7 +315,7 @@ defmodule Membrane.Core.Element.ActionHandler do
          }
          when stream_format != nil <- pad_data do
       state =
-        DemandController.handle_outgoing_buffers(pad_ref, buffers, state)
+        DemandController.decrease_demand_by_outgoing_buffers(pad_ref, buffers, state)
         |> PadModel.set_data!(pad_ref, :start_of_stream?, true)
 
       Message.send(pid, :buffer, buffers, for_pad: other_ref)
